+++ tests/unit/simple_mock_invokemethod_v1old.py
@@ -1,23 +1,26 @@
 """
-mock_pywbem test script that installs a a method provider for the class
-CIM_Foo
+Test mock script that installs a test method provider for CIM method
+method1() in CIM class CIM_Foo, using the old setup approach
+with global variables.
+
+Note: This script and its method provider perform checks because their purpose
+is to test the provider dispatcher. A real mock script with a real method
+provider would not need to perform any of these checks.
 """
 
-from pywbem_mock import MethodProvider
+import pywbem
+import pywbem_mock
 
-from pywbem import CIM_ERR_METHOD_NOT_AVAILABLE, CIMError, CIM_ERR_NOT_FOUND, \
-    CIMInstanceName
-
-# test that GLOBALS exist
 assert "CONN" in globals()
 assert 'SERVER' in globals()
 assert 'VERBOSE' in globals()
 global CONN  # pylint: disable=global-at-module-level
 
 
-class CIM_FooMethodProvider(MethodProvider):
+class CIM_FooMethodProvider(pywbem_mock.MethodProvider):
     """
-    User test provider for InvokeMethod using CIM_Foo and method1.
+    Test method provider for CIM_Foo.method1().
+
     This is basis for testing passing of input parameters correctly and
     generating some exceptions.  It uses only one input parameter where the
     value defines the test and one return parameter that provides data from the
@@ -45,16 +48,16 @@ class CIM_FooMethodProvider(MethodProvider):
 
         # Test if class exists.
         if not self.class_exists(namespace, classname):
-            raise CIMError(
-                CIM_ERR_NOT_FOUND,
+            raise pywbem.CIMError(
+                pywbem.CIM_ERR_NOT_FOUND,
                 "class {0} does not exist in CIM repository, "
                 "namespace {1}".format(classname, namespace))
 
-        if isinstance(localobject, CIMInstanceName):
+        if isinstance(localobject, pywbem.CIMInstanceName):
             instance_store = self.cimrepository.get_instance_store(namespace)
             if not instance_store.object_exists(localobject):
-                raise CIMError(
-                    CIM_ERR_NOT_FOUND,
+                raise pywbem.CIMError(
+                    pywbem.CIM_ERR_NOT_FOUND,
                     "Instance {0} does not exist in CIM repository",
                     format(localobject))
 
@@ -72,11 +75,11 @@ class CIM_FooMethodProvider(MethodProvider):
 
             return (return_value, return_params)
 
-        raise CIMError(CIM_ERR_METHOD_NOT_AVAILABLE)
+        raise pywbem.CIMError(pywbem.CIM_ERR_METHOD_NOT_AVAILABLE)
 
 
-# Add the the callback to the mock repository
+# Register the provider to the mock environment
 # pylint: disable=undefined-variable
-CONN.register_provider(CIM_FooMethodProvider(CONN.cimrepository),  # noqa: F821
-                       CONN.default_namespace,  # noqa: F821
-                       verbose=False)
+_PROV = CIM_FooMethodProvider(CONN.cimrepository)  # noqa: F821
+CONN.register_provider(_PROV, CONN.default_namespace,  # noqa: F821
+                       verbose=VERBOSE)  # noqa: F821